+++ src/main/java/info/u_team/u_team_core/intern/init/UCoreColors.java
@@ -1,9 +1,7 @@
 package info.u_team.u_team_core.intern.init;
 
-import java.util.*;
-
 import info.u_team.u_team_core.UCoreMain;
-import info.u_team.u_team_core.api.IDyeableItem;
+import info.u_team.u_team_core.api.dye.*;
 import net.minecraft.item.Item;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ColorHandlerEvent;
@@ -13,12 +11,6 @@ import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 @EventBusSubscriber(modid = UCoreMain.MODID, value = Dist.CLIENT)
 public class UCoreColors {
 	
-	private static final Set<Item> DYEABLE_ITEMS = new HashSet<>();
-	
-	public static <T extends Item & IDyeableItem> void addItem(T item) {
-		DYEABLE_ITEMS.add(item);
-	}
-	
 	@SubscribeEvent
 	public static void register(ColorHandlerEvent.Item event) {
 		event.getItemColors().register((itemstack, index) -> {
@@ -27,6 +19,6 @@ public class UCoreColors {
 				return ((IDyeableItem) item).getColor(itemstack);
 			}
 			return 0;
-		}, DYEABLE_ITEMS.toArray(new Item[DYEABLE_ITEMS.size()]));
+		}, DyeableItemsRegistry.getDyeableItems().toArray(new Item[DyeableItemsRegistry.getDyeableItems().size()]));
 	}
 }